+++ app/Http/Controllers/CategoryController.php
@@ -5,15 +5,38 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index(){
+        $permissions = [];
+
+        if(Auth::user()->is_super_admin) {
+            $permissions = ['IS_SUPER_ADMIN'];
+        } else if(count(Auth::user()->roles)) {
+            $permissions = Auth::user()->roles[0]->permissions;
+        }
+        if(!user_is_authorized($permissions, 'VIEW_CATEGORY')){
+            return redirect()->back();
+        }
+
         $categories = Category::all();
         return view('categories.index')->with('categories', $categories);
     }
 
     public function create(Request $request){
+        $permissions = [];
+
+        if(Auth::user()->is_super_admin) {
+            $permissions = ['IS_SUPER_ADMIN'];
+        } else if(count(Auth::user()->roles)) {
+            $permissions = Auth::user()->roles[0]->permissions;
+        }
+        if(!user_is_authorized($permissions, 'CREATE_CATEGORY')){
+            return redirect()->back();
+        }
+
         Validator::make($request->all(), [
             'title' => 'required|string|max:255|unique:categories',
             'icon' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
@@ -37,6 +60,17 @@ class CategoryController extends Controller
     }
 
     public function delete(Request $request){
+        $permissions = [];
+
+        if(Auth::user()->is_super_admin) {
+            $permissions = ['IS_SUPER_ADMIN'];
+        } else if(count(Auth::user()->roles)) {
+            $permissions = Auth::user()->roles[0]->permissions;
+        }
+        if(!user_is_authorized($permissions, 'DELETE_CATEGORY')){
+            return redirect()->back();
+        }
+
         $category = Category::find($request->category_id);
 
         if($category->delete()){
